+++ src/components/Form/HorizontalSelect/HorizontalSelect.stories.tsx
@@ -7,13 +7,24 @@
 import React from "react";
 import { Meta, Story } from "@storybook/react/types-6-0";
 import { HorizontalSelect, HorizontalSelectProps } from ".";
+import { SelectOption } from "../../../types";
 
 export default {
   title: "Components/Form/HorizontalSelect",
   component: HorizontalSelect,
 } as Meta;
 
-const Template: Story<HorizontalSelectProps<string>> = (args) => <HorizontalSelect {...args} />;
+const Template: Story<HorizontalSelectProps<string>> = (args) => {
+  const [value, setValue] = React.useState<string>(args.value);
+
+  const onValueChange = (value: string) => setValue(value)
+
+  React.useEffect(() => {
+    setValue(args.value)
+  }, [args.value])
+
+  return <HorizontalSelect {...args} value={value} onChange={onValueChange} />;
+};
 
 export const Example = Template.bind({});
 Example.args = {